Integrate: [ 28759] Silence a bunch of Borland compiler warnings in SDBM_File (See: http://www.nntp.perl.org/group/perl.daily-build.reports/40283)
[ 28786]
Subject: [PATCH perl-current]; was RE: [PATCH perlio.c pp.c regcomp.c] silence format and comparison warnings
From: "Robin Barker" <Robin.Barker@npl.co.uk>
Date: Fri, 1 Sep 2006 16:10:25 +0100
Message-ID: <2C2E01334A940D4792B3E115F95B7226120A21@exchsvr1.npl.ad.local>

[ 28793]
cxtype is a U16, not a U32 (Stops VC++ complaining)

[ 28813]
Subject: NULL and sv_setpv vs. newSVpv
From: "Yitzchak Scott-Thoennes" <sthoenna@efn.org>
Date: Sun, 10 Sep 2006 20:31:46 -0700 (PDT)
Message-ID: <8750.24.19.24.170.1157945506.squirrel@24.19.24.170>

[ 28863]
Fix one genuine unitialised code path spotted by gcc, and a signed/
unsigned warning.

[ 28865]
Subject: [PATCH] bleed -  silence warning in sv.c
From: "Philip M. Gollucci" <pgollucci@p6m7g8.com>
Date: Mon, 18 Sep 2006 01:30:05 -0700
Message-ID: <450E590D.1060205@p6m7g8.com>

[ 28886]
Subject: Re: Change 28877: [PATCH] deal with some gcc warnings
From: Jarkko Hietaniemi <jhi@iki.fi>
Message-ID: <45175CA4.3020900@iki.fi>
Date: Mon, 25 Sep 2006 07:35:48 +0300

remove casts that were causing failure on FreeBSD -Duse64bitint, and
some other cleanups.

[ 28899]
Silence two warnings from gcc when being -pedantic
p4raw-link: @28899 on //depot/perl: dfeee9b153e7ebbeaa1e263dad19a3e5a819bacd
p4raw-link: @28886 on //depot/perl: 6a1e1159a8a738e1607626a2dc2479f546ab4e49
p4raw-link: @28865 on //depot/perl: 1d9c78c6de744b5c7eec4470968743afe4f03acd
p4raw-link: @28863 on //depot/perl: 7f39519f31fe8f3aa4f113c542472c46ec6c535a
p4raw-link: @28813 on //depot/perl: ddfa59c744ee4cca9b8e77d5034cfbacac89bcda
p4raw-link: @28793 on //depot/perl: df43650bac48846555252ca61d5394988cdf564f
p4raw-link: @28786 on //depot/perl: f5992bc4a0a918eda67e6097aac8bd75a3b524e4
p4raw-link: @28759 on //depot/perl: be64226e5b5d608a2b57cf611954f2a7cec0b340

p4raw-id: //depot/maint-5.8/perl@30040
p4raw-integrated: from //depot/perl@28886 'ignore' doio.c reentr.c
	reentr.pl (@28877..)
p4raw-integrated: from //depot/perl@28793 'edit in' pp_ctl.c (@28773..)
p4raw-integrated: from //depot/perl@28786 'edit in' dump.c (@28674..)
	perl.h (@28741..) pp.c (@28771..) sv.c (@28773..) 'merge in'
	pp_pack.c (@28662..) perlio.c utf8.c (@28734..)
p4raw-integrated: from //depot/perl@28759 'merge in' XSUB.h (@28734..)

+++ doio.c
@@ -2218,7 +2218,8 @@ Perl_do_shmio(pTHX_ I32 optype, SV **mark, SV **sp)
     SETERRNO(0,0);
     if (shmctl(id, IPC_STAT, &shmds) == -1)
 	return -1;
-    if (mpos < 0 || msize < 0 || (size_t)mpos + msize > shmds.shm_segsz) {
+    if (mpos < 0 || msize < 0
+	|| (size_t)mpos + msize > (size_t)shmds.shm_segsz) {
 	SETERRNO(EFAULT,SS_ACCVIO);		/* can't do as caller requested */
 	return -1;
     }
